finish

+++ src/main/java/az/matrix/linkedinclone/service/impl/PostServiceImpl.java
@@ -5,15 +5,20 @@ import az.matrix.linkedinclone.dao.entity.Post;
 import az.matrix.linkedinclone.dao.entity.User;
 import az.matrix.linkedinclone.dao.repo.PostRepository;
 import az.matrix.linkedinclone.dto.request.PostRequest;
+import az.matrix.linkedinclone.dto.response.AuthorResponse;
 import az.matrix.linkedinclone.dto.response.PostResponse;
+import az.matrix.linkedinclone.enums.AuthorType;
+import az.matrix.linkedinclone.exception.ForbiddenException;
 import az.matrix.linkedinclone.exception.ResourceNotFoundException;
 import az.matrix.linkedinclone.mapper.PostMapper;
 import az.matrix.linkedinclone.service.PostService;
 import az.matrix.linkedinclone.utility.AuthHelper;
+import az.matrix.linkedinclone.utility.AuthorHelper;
 import az.matrix.linkedinclone.utility.MediaUploadUtil;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,6 +26,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -30,6 +36,7 @@ public class PostServiceImpl implements PostService {
     private final PostMapper postMapper;
     private final PostRepository postRepository;
     private final AuthHelper authHelper;
+    private final AuthorHelper authorHelper;
 
     @Value("${files.directory}")
     private String UPLOAD_DIR;
@@ -37,47 +44,39 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public PostResponse getPost(Long postId) {
-        User user = authHelper.getAuthenticatedUser();
         log.info("Operation of getting post with ID {} started.", postId);
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> {
-                    log.warn("Failed to get post: Post with ID {} not found", postId);
-                    return new ResourceNotFoundException("POST_NOT_FOUND");
-                });
+        Post post = postRepository.findById(postId).orElseThrow(() -> new ResourceNotFoundException(Post.class));
+        PostResponse response = mapPostToResponse(post);
         log.info("Post with ID {} successfully returned", postId);
-        return postMapper.toDto(post);
+        return response;
     }
 
     @Override
-    public Page<PostResponse> getPosts(Pageable pageable) {
+    public Page<PostResponse> getPostsOfProfile(Long authorId, AuthorType authorType, Pageable pageable) {
         log.info("Operation of getting posts started");
-        Page<Post> posts = postRepository.findAll(pageable);
-        Page<PostResponse> postResponsePage = posts.map(postMapper::toDto);
+        Page<Post> posts = postRepository.findAllByAuthorIdAndAuthorType(authorId, authorType, pageable);
+        Page<PostResponse> responsePage = posts.map(this::mapPostToResponse);
         log.info("Posts are returned successfully");
-        return postResponsePage;
+        return responsePage;
     }
 
     @Override
     @Transactional
-    public PostResponse createPost(String text, List<MultipartFile> files) {
+    public PostResponse createPost(PostRequest postRequest, List<MultipartFile> files) {
         User user = authHelper.getAuthenticatedUser();
-        log.info("Operation of creating new post started by user with email {}", user.getEmail());
+        log.info("Creating new post started");
         Post post = new Post();
-        post.setText(text);
-        post.setUser(user);
         if (files != null && !files.isEmpty()) {
-            for (MultipartFile file : files) {
-                String fileUrl = MediaUploadUtil.uploadMediaForPost(file, UPLOAD_DIR);
-                Photo postPhoto = Photo.builder()
-                        .photoPath(fileUrl)
-                        .post(post)
-                        .build();
-                post.getPhotos().add(postPhoto);
-            }
+            handleMediaUpload(files, post);
         }
+        Pair<Long, AuthorType> authorDetails = authorHelper.determineAuthorDetails(user, postRequest.getCompanyId());
+        post.setAuthorId(authorDetails.getLeft());
+        post.setAuthorType(authorDetails.getRight());
+        post.setText(postRequest.getText());
         postRepository.save(post);
-        log.info("Post is successfully created");
-        return postMapper.toDto(post);
+        PostResponse response = mapPostToResponse(post);
+        log.info("Post successfully created");
+        return response;
     }
 
     @Override
@@ -85,15 +84,16 @@ public class PostServiceImpl implements PostService {
     public PostResponse editPost(Long postId, PostRequest postRequest) {
         User user = authHelper.getAuthenticatedUser();
         log.info("Operation of editing post with id {} is started by user with email {}", postId, user.getEmail());
-        Post post = postRepository.findByIdAndUserEmail(postId, user.getEmail())
-                .orElseThrow(() -> {
-                    log.warn("Failed to edit post: Post with ID {} not found or unauthorized access by user with email {}", postId, user.getEmail());
-                    return new ResourceNotFoundException("POST_NOT_FOUND");
-                });
+        Post post = postRepository.findById(postId).orElseThrow(() -> new ResourceNotFoundException(Post.class));
+        if (!authorHelper.validatePermission(post.getAuthorType(), post.getAuthorId(), user)) {
+            log.error("User with ID {} doesn't have permission to edit post with ID {}", user.getId(), postId);
+            throw new ForbiddenException("NOT_ALLOWED");
+        }
         postMapper.mapForUpdate(post, postRequest);
         postRepository.save(post);
+        PostResponse response = mapPostToResponse(post);
         log.info("Post is successfully edited.");
-        return postMapper.toDto(post);
+        return response;
     }
 
     @Override
@@ -101,12 +101,27 @@ public class PostServiceImpl implements PostService {
     public void deletePost(Long postId) {
         User user = authHelper.getAuthenticatedUser();
         log.info("Operation of deleting post with id {} is started by user with email {}", postId, user.getEmail());
-        Post post = postRepository.findByIdAndUserEmail(postId, user.getEmail())
-                .orElseThrow(() -> {
-                    log.warn("Failed to delete post: Post with ID {} not found or unauthorized access by user with email {}", postId, user.getEmail());
-                    return new ResourceNotFoundException("POST_NOT_FOUND");
-                });
+        Post post = postRepository.findById(postId).orElseThrow(() -> new ResourceNotFoundException(Post.class));
+        if (!authorHelper.validatePermission(post.getAuthorType(), post.getAuthorId(), user)) {
+            log.error("User with ID {} doesn't have permission to delete post with ID {}", user.getId(), postId);
+            throw new ForbiddenException("NOT_ALLOWED");
+        }
         postRepository.delete(post);
         log.info("Post is successfully deleted.");
+    }
+
+    private void handleMediaUpload(List<MultipartFile> files, Post post) {
+        for (MultipartFile file : files) {
+            String fileUrl = MediaUploadUtil.uploadMediaForPost(file, UPLOAD_DIR);
+            Photo postPhoto = Photo.builder().photoPath(fileUrl).post(post).build();
+            post.getPhotos().add(postPhoto);
+        }
+    }
+
+    private PostResponse mapPostToResponse(Post post) {
+        PostResponse postResponse = postMapper.toDto(post);
+        AuthorResponse authorResponse = authorHelper.getAuthor(post.getAuthorId(), post.getAuthorType());
+        postResponse.setAuthor(authorResponse);
+        return postResponse;
     }
 }
